% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinker.R
\name{get_categories_shrinker}
\alias{get_categories_shrinker}
\title{List All Web Shrinker Categories}
\usage{
get_categories_shrinker(taxonomy = "iabv1")
}
\arguments{
\item{taxonomy}{character. Which category taxonomy to use, either "iabv1" or "webshrinker"}
}
\value{
list of categories
}
\description{
Returns all of the possible categories from webshrinker.com that URLs, hostnames, and IP addresses can be associated with.
}
\examples{
\dontrun{
get_categories_shrinker(taxonomy = "iabv1")
}
}
