% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klazify.R
\name{get_url_category_klazify}
\alias{get_url_category_klazify}
\title{Klazify Category Lookup}
\usage{
get_url_category_klazify(url, raw = FALSE)
}
\arguments{
\item{url}{character. URL to be categorized (needs https://www.)}

\item{raw}{boolean. If TRUE returns the raw API response which contains more
information than only the categories.}
}
\value{
data.frame with the inferred category and confidents score
}
\description{
uses klazify.com to get the categories associated with the given URL.
}
\examples{
\dontrun{
get_url_category_klazify("https://www.kicker.de")
}
}
