% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkpreview.R
\name{get_url_category_linkpreview}
\alias{get_url_category_linkpreview}
\title{linkpreviewer}
\usage{
get_url_category_linkpreview(url)
}
\arguments{
\item{url}{character. URL to be categorized}
}
\value{
data.frame with the title and description of a website
}
\description{
use linkpreview to get the description of a URL.
}
\examples{
\dontrun{
get_url_category_linkpreview("https://www.google.com")
}
}
