% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoisxml.R
\name{get_url_category_whoisxml}
\alias{get_url_category_whoisxml}
\title{Whoisxml Category Lookup}
\usage{
get_url_category_whoisxml(url)
}
\arguments{
\item{url}{character. URL to be categorized (doesnt need https://www.)}
}
\value{
data.frame with the inferred category and confidents score
}
\description{
uses whoisxmlapi to get the categories associated with the given URL
}
\examples{
\dontrun{
get_url_category_whoisxml("kicker.de")
}
}
