% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{online_news}
\alias{online_news}
\title{Domain list of online news}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 542 rows and 2 columns.
}
\usage{
online_news
}
\description{
Domain list of online news
}
\references{
Sebastian Stier, Nora Kirkizh, Caterina Froio, Ralph Schroeder. Populist
attitudes and selective exposure to online news \url{https://osf.io/5pe27}
}
\keyword{datasets}
